# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/m2m/clients/{apiClient}": {
            "delete": {
                "operationId": "apiClient.destroy",
                "description": "DELETE /api/m2m/clients/{apiClient}\nAuth: auth:api (admin only, same org via ApiClientPolicy::delete)\n\nWrite ordering invariant (design \u00a7Revocation write ordering):\n  1. DB write (is_active=false) committed FIRST \u2014 durable authoritative flag\n  2. Redis denylist write SECOND \u2014 fast-path cache\n\nA crash between the two writes leaves the system in the safer state:\nis_active=false in DB \u2192 next guard lookup rejects the key.",
                "summary": "Revoke an M2M API client",
                "tags": [
                    "ApiClient"
                ],
                "parameters": [
                    {
                        "name": "apiClient",
                        "in": "path",
                        "required": true,
                        "description": "The api client ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/avatar-templates/{id}": {
            "get": {
                "operationId": "avatarTemplate.show",
                "tags": [
                    "AvatarTemplate"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AvatarTemplateResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AvatarTemplateResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "avatarTemplate.update",
                "tags": [
                    "AvatarTemplate"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "description": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "maxLength": 500
                                    },
                                    "config": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AvatarTemplateResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AvatarTemplateResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "avatarTemplate.destroy",
                "tags": [
                    "AvatarTemplate"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "409": {
                        "description": "Deleting what candidates are currently being interviewed with is\na decision, not a cleanup. 409 rather than 422: the request is\nwell-formed, the state is what refuses it.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "const": "template_active"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Activate another template before deleting this one."
                                        }
                                    },
                                    "required": [
                                        "error",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/projects/{project}": {
            "get": {
                "operationId": "projects.show",
                "description": "Returns a single project (TenantScoped \u2014 404 for cross-org requests).\nProject is resolved manually (not via route model binding) to ensure the\nTenantScoped global scope is active (TenantContext runs before this method).",
                "summary": "GET /api/projects/{project}",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ProjectResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProjectResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "projects.update",
                "description": "Updates a project. Immutability and lifecycle guards enforced at FormRequest layer.\nModel guard provides backstop for non-HTTP paths.\nProject is resolved manually \u2014 see class docblock.",
                "summary": "PATCH /api/projects/{project}",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UpdateProjectRequest"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "competency_ids": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProjectResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProjectResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "projects.destroy",
                "description": "Soft-deletes the project. Returns HTTP 204 No Content.\nThe pinned FrameworkVersion remains locked (soft-delete does not unlock).\nProject is resolved manually \u2014 see class docblock.",
                "summary": "DELETE /api/projects/{project}",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AvatarTemplateResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "provider": {
                        "type": "string"
                    },
                    "config": {
                        "type": "array",
                        "items": {}
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description",
                    "provider",
                    "config",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "AvatarTemplateResource"
            },
...
            "ProjectResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "organization_id": {
                        "type": "integer"
                    },
                    "framework_version_id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "assessment_type": {
                        "type": "string"
                    },
                    "role_code": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "language": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "pause_every_n_competencies": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "nudge_min_chars": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "exit_redirect_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "webhook_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "webhook_events": {
                        "type": "array",
                        "items": {}
                    },
                    "deadline_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "webhook_secret intentionally excluded (hidden + encrypted)"
                    },
                    "goes_live_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "pin_context": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Pin context: the FrameworkVersion this project is pinned to",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "version": {
                                "type": "string"
                            },
                            "label": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "is_locked": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "id",
                            "version",
                            "label",
                            "is_locked"
                        ]
                    },
                    "competencies": {
                        "type": "array",
                        "description": "Competencies with position pivot",
                        "items": {}
                    }
                },
                "required": [
                    "id",
                    "organization_id",
                    "framework_version_id",
                    "slug",
                    "name",
                    "assessment_type",
                    "role_code",
                    "language",
                    "status",
                    "pause_every_n_competencies",
                    "nudge_min_chars",
                    "exit_redirect_url",
                    "webhook_url",
                    "webhook_events",
                    "deadline_at",
                    "goes_live_at",
                    "created_at",
                    "updated_at",
                    "pin_context",
                    "competencies"
                ],
                "title": "ProjectResource"
            },
...
            "UpdateProjectRequest": {
                "type": "object",
                "description": "UpdateProjectRequest (C4 Project Configuration).\n\nValidates PATCH /api/projects/{id} payload.\n\nKey invariants:\n- framework_version_id: blanket-prohibited in ALL PATCH requests (immutable from creation).\n  Any PATCH that includes this field is rejected with 422 \u2014 even the same value, even on draft.\n- slug: self-ignoring unique rule (->ignore) with soft-delete exclusion\n- Immutability: changing assessment_type or role_code when the resulting status is 'active' or\n  'archived' \u2192 422\n- Lifecycle: allowed transitions are draft\u2192active and active\u2192archived only.\n  Forbidden (active\u2192draft, archived\u2192active, archived\u2192draft) \u2192 422.\n\nSubstituteBindings note: the route parameter 'project' is an int (no implicit model binding \u2014\nSubstituteBindings runs BEFORE TenantContext in the api middleware group, so route model binding\nwould resolve Project before the tenant scope is set). Manual findOrFail() inside controller\nand FormRequest methods ensures the TenantScoped global scope is active at resolution time.",
                "title": "UpdateProjectRequest"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}